google now requires an API key to access the stats
anonymous access no longer works. we need to use an API key.
we put the API key in a keyFile instead of directly into the config
file/class. this is so that the keyfile can be in an arbitrary
location (ie. non-public, more secure) and not just in the config
file.

for more info on the API key, see
https://developers.google.com/url-shortener/v1/getting_started#APIKey